Position things relative to a fixed point
Up until now, things have been positioned relative to the top-left
corner of the canvas. Doing that meant that positions needed to be
updated if the canvas "scrolled". Now there is a separate coordinate
system that represents positions in the game world, and the world
coordinates are converted to canvas coordinates when rendering.

A camera object represents how much of the world is rendered to the
canvas. When the player moves to specified points on the canvas, the
camera position is updated to make the view "scroll" up or down.

A scene graph is used to store, update, and render entities. Entities
in the scene graph are updated and rendered using a depth-first search.
Map tiles are rendered first due to the order in which they are added
to the graph, then entities.